Layout image is now created with separate class
